refactor: move service packages into the `server/service/` directory
